Split water_pressure.rs (716 lines) into water_pressure/ subdirectory
Breaks the monolithic water_pressure.rs into focused sub-modules:
- types.rs: constants, components, PressureCategory enum, pure helper functions
- state.rs: WaterPressureState resource, Default impl, Saveable impl
- systems.rs: update_water_pressure system and WaterPressurePlugin
- tests.rs: all unit tests
- mod.rs: module declarations and pub re-exports

All public items are re-exported from mod.rs so external callers are unaffected.

Closes #1389
