+++ crates/simulation/src/water_pressure/tests.rs
@@ -1,274 +1,8 @@
-//! Water Pressure Zones and Distribution (WATER-010).
-//!
-//! Higher elevation areas require pumping stations to maintain adequate water
-//! pressure. Buildings above a pressure zone's effective elevation lose water
-//! service or receive reduced service quality.
-//!
-//! - Base pressure zone serves buildings up to elevation 50.
-//! - Booster pump station: extends pressure zone by +30 elevation, $200K, 1x1.
-//! - Buildings above the pressure zone elevation have reduced water pressure
-//!   (lower service quality).
-//! - No pressure = no water service for high-elevation buildings.
-//! - Multiple pump stations can chain (each adds +30 elevation to zone).
-
-use bevy::prelude::*;
-use bitcode::{Decode, Encode};
-use serde::{Deserialize, Serialize};
-
-use crate::buildings::Building;
-use crate::grid::WorldGrid;
-use crate::SlowTickTimer;
-
-// =============================================================================
-// Constants
-// =============================================================================
-
-/// Base elevation limit served by the water distribution system without any
-/// booster pump stations.
-pub const BASE_PRESSURE_ELEVATION: f32 = 50.0;
-
-/// Additional elevation capacity provided by each booster pump station.
-pub const BOOSTER_ELEVATION_GAIN: f32 = 30.0;
-
-/// Cost in dollars to build a booster pump station.
-pub const BOOSTER_PUMP_COST: f64 = 200_000.0;
-
-/// Elevation range over which water pressure degrades from full (1.0) to zero.
-/// Buildings within this margin above the effective pressure elevation receive
-/// reduced service quality (linearly interpolated).
-pub const PRESSURE_FALLOFF_RANGE: f32 = 10.0;
-
-// =============================================================================
-// Components
-// =============================================================================
-
-/// Marker component for booster pump station entities.
-///
-/// Each booster pump station adds `BOOSTER_ELEVATION_GAIN` to the city's
-/// effective pressure elevation. Placed on the grid as a 1x1 building.
-#[derive(Component, Debug, Clone, Serialize, Deserialize)]
-pub struct BoosterPumpStation {
-    /// Grid X position of this pump station.
-    pub grid_x: usize,
-    /// Grid Y position of this pump station.
-    pub grid_y: usize,
-}
-
-// =============================================================================
-// Resource
-// =============================================================================
-
-/// City-wide water pressure zone state.
-///
-/// Tracks the effective pressure elevation (base + booster contributions),
-/// the number of active booster pump stations, and statistics about buildings
-/// served and underserved by the water pressure system.
-#[derive(Resource, Debug, Clone, Serialize, Deserialize, Encode, Decode)]
-pub struct WaterPressureState {
-    /// Number of active booster pump stations.
-    pub booster_count: u32,
-    /// Effective maximum elevation served at full pressure.
-    /// Equal to `BASE_PRESSURE_ELEVATION + booster_count * BOOSTER_ELEVATION_GAIN`.
-    pub effective_elevation: f32,
-    /// Number of buildings with full water pressure (elevation <= effective_elevation).
-    pub buildings_full_pressure: u32,
-    /// Number of buildings with reduced water pressure (in the falloff range).
-    pub buildings_reduced_pressure: u32,
-    /// Number of buildings with no water pressure (above effective_elevation + falloff).
-    pub buildings_no_pressure: u32,
-    /// Average pressure factor across all buildings (0.0 to 1.0).
-    pub average_pressure_factor: f32,
-    /// Total cost of all booster pump stations.
-    pub total_booster_cost: f64,
-}
-
-impl Default for WaterPressureState {
-    fn default() -> Self {
-        Self {
-            booster_count: 0,
-            effective_elevation: BASE_PRESSURE_ELEVATION,
-            buildings_full_pressure: 0,
-            buildings_reduced_pressure: 0,
-            buildings_no_pressure: 0,
-            average_pressure_factor: 1.0,
-            total_booster_cost: 0.0,
-        }
-    }
-}
-
-// =============================================================================
-// Saveable implementation
-// =============================================================================
-
-impl crate::Saveable for WaterPressureState {
-    const SAVE_KEY: &'static str = "water_pressure";
-
-    fn save_to_bytes(&self) -> Option<Vec<u8>> {
-        // Skip saving if at default state (no boosters).
-        if self.booster_count == 0
-            && self.buildings_full_pressure == 0
-            && self.buildings_reduced_pressure == 0
-            && self.buildings_no_pressure == 0
-        {
-            return None;
-        }
-        Some(bitcode::encode(self))
-    }
-
-    fn load_from_bytes(bytes: &[u8]) -> Self {
-        crate::decode_or_warn(Self::SAVE_KEY, bytes)
-    }
-}
-
-// =============================================================================
-// Helper functions (pure, testable)
-// =============================================================================
-
-/// Calculate the effective pressure elevation from the number of booster pump
-/// stations.
-pub fn effective_pressure_elevation(booster_count: u32) -> f32 {
-    BASE_PRESSURE_ELEVATION + booster_count as f32 * BOOSTER_ELEVATION_GAIN
-}
-
-/// Calculate the water pressure factor for a building at a given elevation.
-///
-/// Returns a value between 0.0 and 1.0:
-/// - 1.0 if the building is at or below the effective elevation (full pressure).
-/// - 0.0 if the building is above `effective_elevation + PRESSURE_FALLOFF_RANGE`.
-/// - Linearly interpolated in between.
-pub fn pressure_factor(building_elevation: f32, effective_elev: f32) -> f32 {
-    if building_elevation <= effective_elev {
-        1.0
-    } else {
-        let excess = building_elevation - effective_elev;
-        if excess >= PRESSURE_FALLOFF_RANGE {
-            0.0
-        } else {
-            1.0 - (excess / PRESSURE_FALLOFF_RANGE)
-        }
-    }
-}
-
-/// Classify a pressure factor into one of three categories:
-/// - Full pressure: factor >= 1.0 (or very close due to floating point).
-/// - No pressure: factor <= 0.0.
-/// - Reduced pressure: everything in between.
-pub fn classify_pressure(factor: f32) -> PressureCategory {
-    if factor >= 1.0 - f32::EPSILON {
-        PressureCategory::Full
-    } else if factor <= f32::EPSILON {
-        PressureCategory::None
-    } else {
-        PressureCategory::Reduced
-    }
-}
-
-/// Pressure classification for a building.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum PressureCategory {
-    /// Full water pressure (factor ~1.0).
-    Full,
-    /// Reduced water pressure (0.0 < factor < 1.0).
-    Reduced,
-    /// No water pressure (factor ~0.0).
-    None,
-}
-
-// =============================================================================
-// System
-// =============================================================================
-
-/// System: Recalculate water pressure zone statistics every slow tick.
-///
-/// 1. Counts active booster pump stations.
-/// 2. Computes effective pressure elevation.
-/// 3. Evaluates each building's elevation against the pressure zone.
-/// 4. Updates statistics on the `WaterPressureState` resource.
-pub fn update_water_pressure(
-    timer: Res<SlowTickTimer>,
-    grid: Res<WorldGrid>,
-    boosters: Query<&BoosterPumpStation>,
-    buildings: Query<&Building>,
-    mut pressure_state: ResMut<WaterPressureState>,
-) {
-    if !timer.should_run() {
-        return;
-    }
-
-    // Step 1: Count booster pump stations.
-    let booster_count = boosters.iter().count() as u32;
-
-    // Step 2: Compute effective elevation.
-    let effective_elev = effective_pressure_elevation(booster_count);
-
-    // Step 3: Evaluate each building.
-    let mut full_count: u32 = 0;
-    let mut reduced_count: u32 = 0;
-    let mut no_count: u32 = 0;
-    let mut pressure_sum: f32 = 0.0;
-    let mut building_count: u32 = 0;
-
-    for building in &buildings {
-        let elevation = grid.get(building.grid_x, building.grid_y).elevation;
-        let factor = pressure_factor(elevation, effective_elev);
-
-        match classify_pressure(factor) {
-            PressureCategory::Full => full_count += 1,
-            PressureCategory::Reduced => reduced_count += 1,
-            PressureCategory::None => no_count += 1,
-        }
-
-        pressure_sum += factor;
-        building_count += 1;
-    }
-
-    // Step 4: Update state.
-    pressure_state.booster_count = booster_count;
-    pressure_state.effective_elevation = effective_elev;
-    pressure_state.buildings_full_pressure = full_count;
-    pressure_state.buildings_reduced_pressure = reduced_count;
-    pressure_state.buildings_no_pressure = no_count;
-    pressure_state.total_booster_cost = booster_count as f64 * BOOSTER_PUMP_COST;
-
-    pressure_state.average_pressure_factor = if building_count > 0 {
-        pressure_sum / building_count as f32
-    } else {
-        1.0 // Default to full pressure when there are no buildings.
-    };
-}
-
-// =============================================================================
-// Plugin
-// =============================================================================
-
-pub struct WaterPressurePlugin;
-
-impl Plugin for WaterPressurePlugin {
-    fn build(&self, app: &mut App) {
-        app.init_resource::<WaterPressureState>();
-
-        // Register for save/load via the SaveableRegistry.
-        app.init_resource::<crate::SaveableRegistry>();
-        app.world_mut()
-            .resource_mut::<crate::SaveableRegistry>()
-            .register::<WaterPressureState>();
-
-        app.add_systems(
-            FixedUpdate,
-            update_water_pressure
-                .after(crate::utilities::propagate_utilities)
-                .in_set(crate::SimulationSet::Simulation),
-        );
-    }
-}
-
-// =============================================================================
-// Tests
-// =============================================================================
+//! Tests for the water pressure system.
 
 #[cfg(test)]
 mod tests {
-    use super::*;
+    use crate::water_pressure::*;
 
     // -------------------------------------------------------------------------
     // Constants tests
